revert: hand-made revert of last commit
